Sync- Version 2.0
Prey and Predators are eating and moving about, but death rates and
other parameters need to be played with/ experimented. The predators
are currently not dying by death rate, and only preys are reproducing.
The simulation is lasting about 13,000 timesteps. Which shows, that
even this extreme condition isn’t stabilized yet. Therefore, I just
need to play with parameters.